+++ demo.sh
@@ -40,7 +40,9 @@ while [ $choice -eq 3 ]; do
 done
 
 # We must do a custom build of pysqlite
-wget http://pysqlite.googlecode.com/files/pysqlite-2.6.0.tar.gz
+if [ ! -f pysqlite-2.6.0.tar.gz ]; then
+    wget http://pysqlite.googlecode.com/files/pysqlite-2.6.0.tar.gz
+fi
 tar xzf pysqlite-2.6.0.tar.gz
 cd pysqlite-2.6.0
 echo "[build_ext]\